1. mysite/mysite/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-taggit==5.0.1
Markdown==3.6
psycopg==3.1.18
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
